use 2-step onboarding, some refactoring
