+++ client/src/locales/ja.json
@@ -391,5 +391,6 @@
 	"You've upgraded your account!": "アカウントをアップグレードしました！",
 	"Unlimited usage and premium features are activated.": "無制限の使用機能とプレミアム機能がアクティブになります。",
 	"Let's go": "さあ行こう",
-	"Token limit reached, this answer may be incomplete. To generate a full answer, please reduce the number of tokens used and regenerate.": "到達したトークン制限は、この答えが不完全になる可能性があります。 完全な回答を生成するには、使用して再生するトークンの数を減らしてください。"
+	"Token limit reached, this answer may be incomplete. To generate a full answer, please reduce the number of tokens used and regenerate.": "到達したトークン制限は、この答えが不完全になる可能性があります。 完全な回答を生成するには、使用して再生するトークンの数を減らしてください。",
+	"Let’s get you started with bloop!": "Bloopを始めましょう！"
 }
